+++ src/main/java/com/en_circle/slt/plugin/actions/swank/UninternSymbolAction.java
@@ -1,0 +1,73 @@
+package com.en_circle.slt.plugin.actions.swank;
+
+import com.en_circle.slt.plugin.SltBundle;
+import com.en_circle.slt.plugin.lisp.LispParserUtil;
+import com.en_circle.slt.plugin.lisp.psi.LispSymbol;
+import com.en_circle.slt.plugin.services.lisp.LispEnvironmentService;
+import com.en_circle.slt.plugin.services.lisp.LispEnvironmentService.LispEnvironmentState;
+import com.en_circle.slt.plugin.swank.requests.UninternSymbol;
+import com.intellij.codeInsight.hint.HintManager;
+import com.intellij.openapi.actionSystem.AnActionEvent;
+import com.intellij.openapi.actionSystem.CommonDataKeys;
+import com.intellij.openapi.application.ApplicationManager;
+import com.intellij.openapi.editor.Editor;
+import com.intellij.openapi.editor.ex.EditorEx;
+import com.intellij.openapi.ui.Messages;
+import com.intellij.psi.PsiDocumentManager;
+import com.intellij.psi.PsiElement;
+import com.intellij.psi.PsiFile;
+import com.intellij.psi.util.PsiTreeUtil;
+import org.jetbrains.annotations.NotNull;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.util.Objects;
+
+public class UninternSymbolAction extends SltSwankAction {
+    private static final Logger log = LoggerFactory.getLogger(UninternSymbolAction.class);
+
+    @Override
+    public void actionPerformed(@NotNull AnActionEvent event) {
+        EditorEx editor = (EditorEx) event.getData(CommonDataKeys.EDITOR);
+        PsiElement element = null;
+        if (editor != null) {
+            PsiDocumentManager psiMgr = PsiDocumentManager.getInstance(Objects.requireNonNull(editor.getProject()));
+            psiMgr.commitDocument(editor.getDocument());
+            PsiFile psiFile = psiMgr.getPsiFile(editor.getDocument());
+            if (psiFile != null) {
+                int caretOffset = editor.getExpectedCaretOffset();
+                element = psiFile.findElementAt(caretOffset);
+                while (element != null && !isSymbol(element)) {
+                    element = psiFile.findElementAt(--caretOffset);
+                }
+            }
+
+            if (element == null) {
+                ApplicationManager.getApplication().invokeLater(() -> HintManager.getInstance()
+                        .showErrorHint(editor, SltBundle.message("slt.ui.swank.uninternsymbol.notsymbol")));
+                return;
+            }
+
+            String packageName = LispParserUtil.getPackage(element);
+            String symbolname = element.getText();
+
+            try {
+                LispEnvironmentService.getInstance(editor.getProject())
+                        .sendToLisp(UninternSymbol.uninternSymbol(symbolname, packageName), false);
+            } catch (Exception e) {
+                log.warn(SltBundle.message("slt.error.start"), e);
+                Messages.showErrorDialog(editor.getProject(), e.getMessage(), SltBundle.message("slt.ui.errors.lisp.start"));
+            }
+        }
+    }
+
+    private boolean isSymbol(PsiElement element) {
+        return PsiTreeUtil.getParentOfType(element, LispSymbol.class, false) != null;
+    }
+
+    @Override
+    protected boolean innerActionCheck(Editor editor, PsiFile file) {
+        return LispEnvironmentService.getInstance(file.getProject())
+                .getState() == LispEnvironmentState.READY;
+    }
+}